+++ Session 2/Exercise 2-1/division.h
@@ -1,3 +1,3 @@
 #pragma once
 
-int esw_division(int a, int b);
+float esw_division(int a, int b);
